+++ CHANGELOG.md
@@ -14,16 +14,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 - Renames the *GitLens History* explorer to *GitLens File History* explorer for better clarity
 
+### Removed
+- Removes `gitlens:activeIsTracked`, `gitlens:activeIsBlameable`, `gitlens:activeIsRevision`, and `gitlens:activeHasRemotes` contexts and consolidates them into `gitlens:activeFileStatus` for better performance and UX
+
 ### Fixed
 - Fixes [#436](https://github.com/eamodio/vscode-gitlens/issues/436) - Copy to clipboard not working
 - Fixes [#442](https://github.com/eamodio/vscode-gitlens/issues/442) - GitLens File History fails if name (or path) starts with -
 - Fixes [#440](https://github.com/eamodio/vscode-gitlens/issues/436) - Searching for commits with an empty query yields to no results anymore
+- Fixes issue where commands in the editor toolbar would flash unnecessarily when navigating history or switching tabs
 - Fixes issue where the *Compare File with Previous Revision* command wouldn't work properly when the file had been renamed in some cases
 - Fixes issue where changed files count was wrong when the branch was behind the upstream
 - Fixes issue where the *GitLens File History* explorer wasn't being updated automatically for working changes
 - Fixes issue where the *Compare File with * Revision* commands in the editor toolbar would show and hide too often because of insignificant focus changes
 - Fixes issue where the scope box would be empty when there was no workspace open in the interactive settings editor
-
 
 ## [8.4.1] - 2018-06-19
 ### Fixed